+++ apps/extension/src/ui/apps/dashboard/layout/notifications/DashboardNotificationsAndModals.tsx
@@ -13,6 +13,7 @@ import { SeekBenefitsModal } from "@ui/domains/Portfolio/SeekBenefits/SeekBenefi
 import { RampsModal } from "@ui/domains/Ramps/RampsModal"
 import { MigratePasswordModal } from "@ui/domains/Settings/MigratePassword/MigratePasswordModal"
 import { BittensorBondModal } from "@ui/domains/Staking/Bittensor/BittensorBondModal"
+import { BittensorClaimSettingsModal } from "@ui/domains/Staking/Bittensor/BittensorClaimSettingsModal"
 import { BondModal } from "@ui/domains/Staking/Bond/BondModal"
 import { NomPoolWithdrawModal } from "@ui/domains/Staking/NomPoolWithdraw/NomPoolWithdrawModal"
 import { UnbondModal } from "@ui/domains/Staking/Unbond/UnbondModal"
@@ -50,6 +51,7 @@ export const DashboardNotificationsAndModals = () => {
       <BondModal />
       <BittensorBondModal />
       <ProductSelectionModal />
+      <BittensorClaimSettingsModal />
       <CopyAddressModal />
       <ExplorerNetworkPickerModal />
       <GetStartedModals />