+++ apps/extension/src/ui/domains/Staking/Bittensor/BittensorBondModal/Forms/index.tsx
@@ -1,5 +1,5 @@
 import { useBittensorBondWizard } from "../../hooks/useBittensorBondWizard"
-import { BittensorBondFollowUp } from "./BittensorBondFollowUp"
+import { BittensorBondFollowUp } from "../BittensorBondFollowUp"
 import { BittensorRootBondForm } from "./BittensorRootBondForm"
 import { BittensorRootBondReview } from "./BittensorRootBondReview"
 import { BittensorStakingPositionSelect } from "./BittensorStakingPositionSelect"